+++ concordium-consensus/src/Concordium/GlobalState/BakerInfo.hs
@@ -220,6 +220,7 @@ data ValidatorAdd = ValidatorAdd
       vaMetadataURL :: !UrlText,
       -- | The commission rates for the validator.
       vaCommissionRates :: !CommissionRates
+      -- TODO (drsk) Github issue #1246. Support suspend/resume for ValidatorAdd.
     }
     deriving (Eq, Show)
 
@@ -241,7 +242,9 @@ data ValidatorUpdate = ValidatorUpdate
       -- | The new baking reward commission for the validator.
       vuBakingRewardCommission :: !(Maybe AmountFraction),
       -- | The new finalization reward commission for the validator.
-      vuFinalizationRewardCommission :: !(Maybe AmountFraction)
+      vuFinalizationRewardCommission :: !(Maybe AmountFraction),
+      -- | Whether the validator's account should be suspended/resumed.
+      vuSuspend :: !(Maybe Bool)
     }
     deriving (Eq, Show)
 
@@ -256,7 +259,8 @@ validatorRemove =
           vuMetadataURL = Nothing,
           vuTransactionFeeCommission = Nothing,
           vuBakingRewardCommission = Nothing,
-          vuFinalizationRewardCommission = Nothing
+          vuFinalizationRewardCommission = Nothing,
+          vuSuspend = Nothing
         }
 
 -- | Failure modes when configuring a validator.
@@ -287,6 +291,8 @@ data BakerConfigureUpdateChange
     | BakerConfigureTransactionFeeCommission !AmountFraction
     | BakerConfigureBakingRewardCommission !AmountFraction
     | BakerConfigureFinalizationRewardCommission !AmountFraction
+    | BakerConfigureSuspended
+    | BakerConfigureResumed
     deriving (Eq, Show)
 
 -- | Parameters for adding a delegator.
@@ -368,6 +374,7 @@ genesisBakerInfoEx spv cp GenesisBaker{..} = case spv of
     SP5 -> binfoV1
     SP6 -> binfoV1
     SP7 -> binfoV1
+    SP8 -> binfoV1
   where
     bkrInfo =
         BakerInfo
@@ -379,9 +386,12 @@ genesisBakerInfoEx spv cp GenesisBaker{..} = case spv of
     binfoV1 :: (PVSupportsDelegation pv, PoolParametersVersionFor (ChainParametersVersionFor pv) ~ 'PoolParametersVersion1) => BakerInfoEx (AccountVersionFor pv)
     binfoV1 =
         BakerInfoExV1
-            bkrInfo
-            BakerPoolInfo
-                { _poolOpenStatus = OpenForAll,
-                  _poolMetadataUrl = emptyUrlText,
-                  _poolCommissionRates = cp ^. cpPoolParameters . ppCommissionBounds . to maximumCommissionRates
-                }
+            { _bieBakerInfo = bkrInfo,
+              _bieBakerPoolInfo =
+                BakerPoolInfo
+                    { _poolOpenStatus = OpenForAll,
+                      _poolMetadataUrl = emptyUrlText,
+                      _poolCommissionRates = cp ^. cpPoolParameters . ppCommissionBounds . to maximumCommissionRates
+                    },
+              _bieAccountIsSuspended = conditionally (sSupportsValidatorSuspension (sAccountVersionFor spv)) False
+            }