+++ concordium-consensus/src/Concordium/GlobalState/Block.hs
@@ -137,6 +137,7 @@ blockVersion SP4 = 2
 blockVersion SP5 = 2
 blockVersion SP6 = 3
 blockVersion SP7 = 3
+blockVersion SP8 = 3
 {-# INLINE blockVersion #-}
 
 -- | Type class that supports serialization of a block.