+++ concordium-consensus/tests/scheduler/SchedulerTests/ConfigureBaker.hs
@@ -42,6 +42,7 @@ import Test.Hspec
 
 -- | Deterministically generate a baker account from a seed.
 makeTestBakerV1FromSeed ::
+    forall av m.
     (IsAccountVersion av, Blob.MonadBlobStore m, AVSupportsDelegation av) =>
     -- | The initial balance of the account.
     Amount ->
@@ -53,14 +54,17 @@ makeTestBakerV1FromSeed ::
     BakerId ->
     -- | Seed used to generate account and baker keys.
     Int ->
+    -- | Whether to suspend the account initially.
+    Bool ->
     m (BS.PersistentAccount av)
-makeTestBakerV1FromSeed amount stake bakerId seed = do
+makeTestBakerV1FromSeed amount stake bakerId seed suspend = do
     account <- Helpers.makeTestAccountFromSeed amount seed
     let (fulBaker, _, _, _) = mkFullBaker seed bakerId
     let bakerInfoEx =
             BakerInfoExV1
                 { _bieBakerInfo = fulBaker ^. theBakerInfo,
-                  _bieBakerPoolInfo = poolInfo
+                  _bieBakerPoolInfo = poolInfo,
+                  _bieAccountIsSuspended = conditionally (sSupportsValidatorSuspension (accountVersion @av)) suspend
                 }
     BS.addAccountBakerV1 bakerInfoEx stake True account
   where
@@ -96,7 +100,7 @@ makeTestDelegatorFromSeed amount accountDelegation seed = do
 baker0Account ::
     (IsAccountVersion av, Blob.MonadBlobStore m, AVSupportsDelegation av) =>
     m (BS.PersistentAccount av)
-baker0Account = makeTestBakerV1FromSeed 1_000_000_000_000 1_000_000_000_000 bakerId seed
+baker0Account = makeTestBakerV1FromSeed 1_000_000_000_000 1_000_000_000_000 bakerId seed False
   where
     bakerId = 0
     seed = 16
@@ -128,7 +132,7 @@ delegator1KP = Helpers.keyPairFromSeed 17
 baker2Account ::
     (IsAccountVersion av, Blob.MonadBlobStore m, AVSupportsDelegation av) =>
     m (BS.PersistentAccount av)
-baker2Account = makeTestBakerV1FromSeed balance stake bakerId seed
+baker2Account = makeTestBakerV1FromSeed balance stake bakerId seed False
   where
     balance = 1_000_000_000_000
     stake = 1_000_000_000
@@ -153,7 +157,16 @@ dummy3KP = Helpers.keyPairFromSeed 19
 baker4Account ::
     (IsAccountVersion av, Blob.MonadBlobStore m, AVSupportsDelegation av) =>
     m (BS.PersistentAccount av)
-baker4Account = makeTestBakerV1FromSeed 20_000_000_000_000 500_000_000_000 bakerId seed
+baker4Account = makeTestBakerV1FromSeed 20_000_000_000_000 500_000_000_000 bakerId seed False
+  where
+    bakerId = 4
+    seed = 20
+
+-- | Account of the baker 5. This account is suspended.
+baker5Account ::
+    (IsAccountVersion av, Blob.MonadBlobStore m, AVSupportsDelegation av) =>
+    m (BS.PersistentAccount av)
+baker5Account = makeTestBakerV1FromSeed 20_000_000_000_000 500_000_000_000 bakerId seed True
   where
     bakerId = 4
     seed = 20
@@ -221,7 +234,8 @@ testDelegatorToBakerOk spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader delegator1Address 1 10_000,
                       keys = [(0, [(0, delegator1KP)])]
@@ -322,7 +336,8 @@ testDelegatorToBakerOk spv pvString =
                               _bakingCommission = makeAmountFraction 1_000,
                               _transactionCommission = makeAmountFraction 1_000
                             }
-                    }
+                    },
+              _bieAccountIsSuspended = conditionally (sSupportsValidatorSuspension (sAccountVersionFor spv)) False
             }
     updateStaking keysWithProofs = case sSupportsFlexibleCooldown (sAccountVersionFor spv) of
         SFalse -> id
@@ -363,7 +378,8 @@ testDelegatorToBakerDuplicateKey spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader delegator1Address 1 10_000,
                       keys = [(0, [(0, delegator1KP)])]
@@ -426,7 +442,8 @@ testDelegatorToBakerMissingParam spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader delegator1Address 1 10_000,
                       keys = [(0, [(0, delegator1KP)])]
@@ -471,7 +488,7 @@ testAddBakerOk ::
     SProtocolVersion pv ->
     String ->
     Spec
-testAddBakerOk _spv pvString =
+testAddBakerOk spv pvString =
     specify (pvString ++ ": AddBaker (OK)") $ do
         keysWithProofs <- makeBakerKeysWithProofs dummy3Address 3
         let transactions =
@@ -485,7 +502,8 @@ testAddBakerOk _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader dummy3Address 1 transactionEnergy,
                       keys = [(0, [(0, dummy3KP)])]
@@ -584,7 +602,8 @@ testAddBakerOk _spv pvString =
                               _bakingCommission = makeAmountFraction 1_000,
                               _transactionCommission = makeAmountFraction 1_000
                             }
-                    }
+                    },
+              _bieAccountIsSuspended = conditionally (sSupportsValidatorSuspension (sAccountVersionFor spv)) False
             }
     updateStaking keysWithProofs =
         ( Transient.accountStaking
@@ -619,7 +638,8 @@ testAddBakerInsufficientBalance _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader dummy3Address 1 transactionEnergy,
                       keys = [(0, [(0, dummy3KP)])]
@@ -677,7 +697,8 @@ testAddBakerMissingParam _spv pvString =
                               cbMetadataURL = Nothing,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader dummy3Address 1 transactionEnergy,
                       keys = [(0, [(0, dummy3KP)])]
@@ -736,7 +757,8 @@ testAddBakerInvalidProofs _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader dummy3Address 1 transactionEnergy,
                       keys = [(0, [(0, dummy3KP)])]
@@ -792,7 +814,8 @@ testUpdateBakerOk _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Nothing,
-                              cbFinalizationRewardCommission = Nothing
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -856,7 +879,8 @@ testUpdateBakerOk _spv pvString =
                 )
                 updatedAccount1
     updateStaking =
-        ( Transient.accountStaking . accountBaker
+        ( Transient.accountStaking
+            . accountBaker
             %~ (stakedAmount .~ stakeAmount)
                 . (stakeEarnings .~ True)
                 . ( accountBakerInfo . bieBakerPoolInfo
@@ -887,7 +911,8 @@ testUpdateBakerInsufficientBalance _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Nothing,
-                              cbFinalizationRewardCommission = Nothing
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -943,7 +968,8 @@ testUpdateBakerLowStake _spv pvString =
                               cbMetadataURL = Just emptyUrlText,
                               cbTransactionFeeCommission = Just (makeAmountFraction 1_000),
                               cbBakingRewardCommission = Nothing,
-                              cbFinalizationRewardCommission = Nothing
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -1001,7 +1027,8 @@ testUpdateBakerInvalidProofs _spv pvString =
                               cbMetadataURL = Nothing,
                               cbTransactionFeeCommission = Nothing,
                               cbBakingRewardCommission = Nothing,
-                              cbFinalizationRewardCommission = Nothing
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -1055,7 +1082,8 @@ testUpdateBakerRemoveOk spv pvString =
                               cbMetadataURL = Nothing,
                               cbTransactionFeeCommission = Nothing,
                               cbBakingRewardCommission = Nothing,
-                              cbFinalizationRewardCommission = Nothing
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -1126,7 +1154,8 @@ testUpdateBakerReduceStakeOk spv pvString =
                               cbMetadataURL = Nothing,
                               cbTransactionFeeCommission = Nothing,
                               cbBakingRewardCommission = Just (makeAmountFraction 1_000),
-                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000)
+                              cbFinalizationRewardCommission = Just (makeAmountFraction 1_000),
+                              cbSuspend = Nothing
                             },
                       metadata = makeDummyHeader baker4Address 1 transactionEnergy,
                       keys = [(0, [(0, baker4KP)])]
@@ -1207,6 +1236,84 @@ testUpdateBakerReduceStakeOk spv pvString =
     accountBaker f (AccountStakeBaker b) = AccountStakeBaker <$> f b
     accountBaker _ x = pure x
 
+data TestSuspendOrResume
+    = Suspend
+    | Resume
+    deriving (Show, Eq)
+
+testUpdateBakerSuspendResumeOk ::
+    forall m pv.
+    (IsProtocolVersion pv, PVSupportsDelegation pv, m ~ Helpers.PersistentBSM pv) =>
+    SProtocolVersion pv ->
+    String ->
+    TestSuspendOrResume ->
+    m (BS.PersistentAccount (AccountVersionFor pv)) ->
+    Spec
+testUpdateBakerSuspendResumeOk spv pvString suspendOrResume accM =
+    specify (pvString ++ ": UpdateBaker: " ++ show suspendOrResume ++ " (OK)") $ do
+        let transactions =
+                [ Runner.TJSON
+                    { payload =
+                        Runner.ConfigureBaker
+                            { cbCapital = Nothing,
+                              cbRestakeEarnings = Nothing,
+                              cbOpenForDelegation = Nothing,
+                              cbKeysWithProofs = Nothing,
+                              cbMetadataURL = Nothing,
+                              cbTransactionFeeCommission = Nothing,
+                              cbBakingRewardCommission = Nothing,
+                              cbFinalizationRewardCommission = Nothing,
+                              cbSuspend = Just $ suspendOrResume == Suspend
+                            },
+                      metadata = makeDummyHeader baker4Address 1 transactionEnergy,
+                      keys = [(0, [(0, baker4KP)])]
+                    }
+                ]
+        (result, doBlockStateAssertions) <-
+            Helpers.runSchedulerTestTransactionJson
+                Helpers.defaultTestConfig
+                (initialBlockState2 @pv)
+                (Helpers.checkReloadCheck checkState)
+                transactions
+        Helpers.assertSuccessWithEvents events result
+        doBlockStateAssertions
+  where
+    -- Transaction length is 64 bytes (4 bytes for the transaction and 60 bytes for the header).
+    transactionEnergy = Cost.configureBakerCostWithoutKeys + Cost.baseCost 64 1
+    checkState ::
+        Helpers.SchedulerResult ->
+        BS.PersistentBlockState pv ->
+        Helpers.PersistentBSM pv Assertion
+    checkState result blockState = do
+        invariants <- Helpers.assertBlockStateInvariantsH blockState (Helpers.srExecutionCosts result)
+        updatedAccount1 <- BS.toTransientAccount . fromJust =<< BS.bsoGetAccountByIndex blockState 4
+        initialAccount1 <- BS.toTransientAccount =<< accM
+        return $ do
+            invariants
+            assertEqual
+                "Expected account update"
+                ( initialAccount1
+                    & Transient.accountNonce .~ 2
+                    & Transient.accountAmount -~ Helpers.energyToAmount transactionEnergy
+                    & updateResumed
+                )
+                updatedAccount1
+    updateResumed = case sSupportsValidatorSuspension (sAccountVersionFor spv) of
+        STrue ->
+            Transient.accountStaking
+                . accountBaker
+                . accountBakerInfo
+                . bieAccountIsSuspended
+                .~ (suspendOrResume == Suspend)
+        SFalse -> id
+    accountBaker f (AccountStakeBaker b) = AccountStakeBaker <$> f b
+    accountBaker _ x = pure x
+    events =
+        [ if suspendOrResume == Suspend
+            then BakerSuspended{ebsBakerId = 4}
+            else BakerResumed{ebrBakerId = 4}
+        ]
+
 tests :: Spec
 tests =
     describe "ConfigureBaker transactions" $
@@ -1214,7 +1321,7 @@ tests =
             Helpers.forEveryProtocolVersion testCases
   where
     testCases :: forall pv. (IsProtocolVersion pv) => SProtocolVersion pv -> String -> Spec
-    testCases spv pvString =
+    testCases spv pvString = do
         case delegationSupport @(AccountVersionFor pv) of
             SAVDelegationNotSupported -> return ()
             SAVDelegationSupported -> do
@@ -1231,3 +1338,8 @@ tests =
                 testUpdateBakerInvalidProofs spv pvString
                 testUpdateBakerRemoveOk spv pvString
                 testUpdateBakerReduceStakeOk spv pvString
+                case sSupportsValidatorSuspension (sAccountVersionFor spv) of
+                    STrue -> do
+                        testUpdateBakerSuspendResumeOk spv pvString Suspend (baker4Account @(AccountVersionFor pv))
+                        testUpdateBakerSuspendResumeOk spv pvString Resume (baker5Account @(AccountVersionFor pv))
+                    SFalse -> return ()